requirements.txt
langchain===0.1.7
langchain-openai==0.0.5
openai==1.12.0
python-dotenv==1.0.0
tenacity==8.1.0
tiktoken==0.5.2
lark==1.1.5
streamlit==1.13.0
altair<5
faiss-cpu==1.7.4
chromadb==0.4.8
pydantic==1.10.8
typing-inspect==0.8.0
typing_extensions
lark==1.1.5
beautifulsoup4==4.12.2
soupsieve==2.5
boto3==1.29.2
botocore==1.32.2
pdfplumber==0.11.0
lxml==5.1.0
python-docx==1.1.0
markdown2==2.4.12
joblib==1.3.2
nltk==3.8.1
networkx==3.2.1
scipy==1.12.0
pysqlite3-binary
